Changes to be committed: 	modified:   IHErrorFile.txt 	modified:   ReadMe.txt 	new file:   bin/tftp/IHErrorFile.txt 	deleted:    bin/tftp/IntermediateHost$1.class 	new file:   bin/tftp/IntermediateHost$1ClientPortHandler.class 	new file:   bin/tftp/IntermediateHost$1ServerPortHandler.class 	modified:   bin/tftp/IntermediateHost$1WellKnownPortHandler.class 	deleted:    bin/tftp/IntermediateHost$2.class 	modified:   bin/tftp/IntermediateHost.class 	modified:   bin/tftp/SendReceiveInterface.class 	modified:   src/tftp/IntermediateHost.java 	modified:   src/tftp/SendReceiveInterface.java
PacketFX has been implemented into intermediateHost.
The three anonymous classes that originally were used have now been changed, they are complete (inner) classes that extend Thread and implement SendReceiveInterface
These three classes contain a lot of code duplication, lots of the things they do are the same and the methods implemented from SendReceiveInterface are all identical. When refactoring, a superclass that extends Thread and implements SendReceiveInterface should be made, then these three classes will extend that new class.

I can't test the entire thing on my linux machine because it doesn't want to let me use some of the ports. However, what has been changed and not tested is fairly trivial code.
The IntermediateHost generatePacketFX() method correctly parses the file, the syntax of the language is in the Readme.txt, and the comment delimiter can easily be changed at the top of IntermediateHost, it is a constant.

Currently, IHErrorFile is in bin/tftp/IHErrorFile.txt, there is a symbolic link to this file in the root directory, but I don't know how that will work on Windows. I can't seem to get Java to pick up the file in ../../IHErrorFile.txt, not sure why.